contract endpoints

+++ api/service_endpoints.py
@@ -21,6 +21,7 @@ class ServiceCreate(BaseModel):
     name: str
     description: str
     service_category_id: str
+    locations: str = None
     sub_category: str =None
     image_paths: str = None
     video_paths: str = None
@@ -31,11 +32,13 @@ class ServiceUpdate(BaseModel):
     description: Optional[str] = None
     service_category_id: Optional[str] = None
     sub_category: Optional[str] = None
+    locations: Optional[str] = None
     image_paths: Optional[str] = None
     video_paths: Optional[str] = None
     banner_paths: Optional[str] = None
 
 class PricePackageCreate(BaseModel):
+    service_id: str
     name: str
     description: str
     price: float
@@ -49,17 +52,22 @@ class PricePackageUpdate(BaseModel):
 
 #create service category
 @service_router.post("/service_category", response_model=dict)
-def create_service_category(category_data: ServiceCategoryCreate, user_id: str = None):
-    if user_id is None:
-        raise HTTPException(status_code=401, detail="User not logged in")
-    user = storage.find_by(User, **{"id": user_id})
+def create_service_category(category_data: ServiceCategoryCreate, session_id: str = Cookie(None)):
+    if session_id is None:
+        raise HTTPException(status_code=403, detail="User not logged in")
+    user = storage.find_by(User, **{"session_id": session_id})
+
     if user is None:
-        raise HTTPException(status_code=401, detail="User not found")
+        raise HTTPException(status_code=403, detail="User session expired")
+
+    if user.to_dict()["is_admin"] != False:
+        raise HTTPException(status_code=403, detail="Unauthorized")
+
     try:
         category_data_dict = category_data.dict()
         if storage.find_by(ServiceCategory, **{"name": category_data_dict['name']}):
             raise HTTPException(status_code=409, detail="Service_Category already exists")
-        category_data_dict['user_id'] = user_id
+
         new_category = ServiceCategory(**category_data.dict())
         storage.new(new_category)
         storage.save()
@@ -68,7 +76,7 @@ def create_service_category(category_data: ServiceCategoryCreate, user_id: str =
         raise HTTPException(status_code=409, detail="Category already exixts")
 
 #Get list of service categories
-@service_router.get("/service_category", response_model=list)
+@service_router.get("/service_categories", response_model=list)
 async def get_service_categories():
     service_categories = storage.all(ServiceCategory).values()
     categories = []
@@ -87,20 +95,29 @@ async def get_service_categories():
 async def get_service_category(id: str):
     service_category = storage.find_by(ServiceCategory, **{"id": id})
     if service_category is None:
-        raise HTTPException(status_code=404, detail="Service category not found")
+        raise HTTPException(status_code=409, detail="Service category not found")
     return service_category.to_dict()
 
 # Update a service category
 @service_router.put("/service_category/{id}", response_model=dict)
 async def update_service_category(id: str, category_data: ServiceCategoryUpdate, session_id: str = Cookie(None)):
+    if session_id is None:
+        raise HTTPException(status_code=403, detail="User not logged in")
+
+    user = storage.find_by(User, **{"session_id": session_id})
+    if user is None:
+        raise HTTPException(status_code=403, detail="User session expired")
+
+    if user.to_dict()["is_admin"] != False:
+        raise HTTPException(status_code=403, detail="Unauthorized")
+
     service_category = storage.find_by(ServiceCategory, **{"id": id})
     if service_category is None:
-        raise HTTPException(status_code=404, detail="Service category not found")
+        raise HTTPException(status_code=409, detail="Service category not found")
 
     try:
         category_data_dict = category_data.dict()
-        for key, value in category_data_dict.items():
-            setattr(service_category, key, value)
+        storage.update(service_category, **category_data_dict)
         storage.save()
         return {"message": "Service category updated successfully", "service_category": service_category.to_dict()}
     except Exception as e:
@@ -108,7 +125,17 @@ async def update_service_category(id: str, category_data: ServiceCategoryUpdate,
 
 # Delete a service category
 @service_router.delete("/service_category/{id}", response_model=dict)
-async def delete_service_category(id: str):
+async def delete_service_category(id: str, session_id: str = Cookie(None)):
+    if session_id is None:
+        raise HTTPException(status_code=403, detail="User not logged in")
+
+    user = storage.find_by(User, **{"session_id": session_id})
+    if user is None:
+        raise HTTPException(status_code=403, detail="User session expired")
+
+    if user.to_dict()["is_admin"] != False:
+        raise HTTPException(status_code=403, detail="Unauthorized")
+        
     service_category = storage.find_by(ServiceCategory, **{"id": id})
     if service_category is None:
         raise HTTPException(status_code=404, detail="Service category not found")
@@ -120,17 +147,20 @@ async def delete_service_category(id: str):
 # Create a service
 @service_router.post("/services/", response_model=dict)
 async def create_service(service_data: ServiceCreate, session_id: str = Cookie(None)):
-    if service_data.user_id is None:
-        raise HTTPException(status_code=401, detail="Invalid User credentials")
+    if session_id is None:
+        raise HTTPException(status_code=403, detail="User not logged in")
 
-    if storage.find_by(User, **{"id": service_data.user_id}) is None:
-        raise HTTPException(status_code=401, detail="Invalid User credentials")
+    user = storage.find_by(User, **{"session_id": session_id})
+    if user is None:
+        raise HTTPException(status_code=403, detail="User session expired")
+        
 
     if storage.find_by(ServiceCategory, **{"id": service_data.service_category_id}) is None:
         raise HTTPException(status_code=401, detail="Invalid Service Category")
 
     try:
         service_data_dict = service_data.dict()
+        service_data_dict["user_id"] =   user.to_dict()["id"]
         service_data_dict["is_verified"] = False
         service = Service(**service_data_dict)
         storage.new(service)
@@ -139,7 +169,7 @@ async def create_service(service_data: ServiceCreate, session_id: str = Cookie(N
     except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
 
-# Get all services
+# Get a list all services
 @service_router.get("/services/", response_model=list)
 async def get_services():
     services = storage.all(Service).values()
@@ -170,12 +200,22 @@ async def get_service(id: str):
 # Update a service
 @service_router.put("/services/{id}", response_model=dict)
 async def update_service(id: str, service_data: ServiceUpdate, session_id: str = Cookie(None)):
+    if session_id is None:
+        raise HTTPException(status_code=403, detail="User not logged in")
+
+    user = storage.find_by(User, **{"session_id": session_id})
+    if user is None:
+        raise HTTPException(status_code=403, detail="User session expired")
+
     service = storage.find_by(Service, **{"id": id})
     if service is None:
         raise HTTPException(status_code=404, detail="Service not found")
 
     try:
         service_data_dict = service_data.dict()
+        if storage.find_by(ServiceCategory, **{"id": service_data_dict['service_category_id']}) is None:
+            raise HTTPException(status_code=409, detail="Invalid Service Category")
+
         storage.update(service, **service_data_dict)
         storage.save()
         service = storage.find_by(Service, **{"id": id})
@@ -185,22 +225,45 @@ async def update_service(id: str, service_data: ServiceUpdate, session_id: str =
 
 # Delete a service
 @service_router.delete("/services/{id}", response_model=dict)
-async def delete_service(id: str):
+async def delete_service(id: str, session_id: str = Cookie(None)):
+    if session_id is None:
+        raise HTTPException(status_code=403, detail="User not logged in")
+    
+    user = storage.find_by(User, **{"session_id": session_id})
+    if user is None:
+        raise HTTPException(status_code=403, detail="User session expired")
+
     service = storage.find_by(Service, **{"id": id})
     if service is None:
         raise HTTPException(status_code=404, detail="Service not found")
+    
+    if service.user_id != user.id:
+        raise HTTPException(status_code=403, detail="Unauthorized")
+
     storage.delete(service)
     storage.save()
     return {"message": "Service deleted successfully"}
 
 # Create a service price package
 @service_router.post("/service_price_package", response_model=dict)
-async def create_service_price_package(package_data: PricePackageCreate, service_id: str):
-    service = storage.find_by(Service, **{"id": service_id})
+async def create_service_price_package(package_data: PricePackageCreate, session_id: str = Cookie(None)):
+    if session_id is None:
+        raise HTTPException(status_code=403, detail="User not logged in")
+    
+    user = storage.find_by(User, **{"session_id": session_id})
+    if user is None:
+        raise HTTPException(status_code=403, detail="User session expired")
+
+    package_data_dict = package_data.dict()
+
+    service = storage.find_by(Service, **{"id": package_data_dict['service_id']})
     if service is None:
         raise HTTPException(status_code=404, detail="Service not found")
+    
+    if service.user_id != user.id:
+        raise HTTPException(status_code=403, detail="Unauthorized")
+    
     try:
-        package_data_dict = package_data.dict()
         package_data_dict["service_id"] = service_id
         price_package = PricePackage(**package_data_dict)
         storage.new(price_package)
@@ -243,10 +306,8 @@ async def update_service_price_package(id: str, package_data: PricePackageUpdate
         raise HTTPException(status_code=404, detail="Service price package not found")
     
     try:
-        package_data_dict = package_data.dict(exclude_unset=True)
-        for key, value in package_data_dict.items():
-            setattr(price_package, key, value)
-        
+        package_data_dict = package_data.dict()
+        storage.update(price_package, **package_data_dict)
         storage.save()
         return {"message": "Service price package updated successfully", "price_package": price_package.to_dict()}
     except Exception as e:
